Break selector for domainData
